Moving chatClient to separate file
