--- jsconfig.json
@@ -1,3 +1,0 @@
-{
-	"exclude": ["node_modules", "**/node_modules/*"]
-}
